+++ tests/Unit/CustomTest.php
@@ -112,6 +112,7 @@ class CustomTest extends TestCase
 		$sequence = 0;
 		
 		$factory = $this->getBitsFactory(7, new TestingSequenceResolver($sequence));
+		$factory->setTestNow(now());
 		
 		$bits_at_epoch1 = $factory->make();
 		
@@ -127,7 +128,7 @@ class CustomTest extends TestCase
 		$this->assertEquals($bits_at_epoch2->values[1], 7);
 		$this->assertEquals($bits_at_epoch2->values[2], 1);
 		
-		Date::setTestNow(now()->addMicrosecond());
+		$factory->setTestNow(now()->addMicrosecond());
 		$bits_at_1us = $factory->make();
 		
 		$this->assertEquals($bits_at_1us->id(), 0b0000000000000000000000000000000000000000000000000101110000000010);
@@ -135,7 +136,7 @@ class CustomTest extends TestCase
 		$this->assertEquals($bits_at_1us->values[1], 7);
 		$this->assertEquals($bits_at_1us->values[2], 2);
 		
-		Date::setTestNow(now()->addMicrosecond());
+		$factory->setTestNow(now()->addMicroseconds(2));
 		$bits_at_2us = $factory->make();
 		
 		$this->assertEquals($bits_at_2us->id(), 0b0000000000000000000000000000000000000000000000001001110000000011);
@@ -174,7 +175,7 @@ class CustomTest extends TestCase
 		
 		$this->assertEquals(0, $first_after_second_sleep->values[2]);
 		
-		Sleep::assertSlept(fn(CarbonInterval $interval) => $interval->totalMicroseconds === 1, 2);
+		Sleep::assertSlept(fn(CarbonInterval $interval) => (int) $interval->totalMicroseconds === 1, 2);
 	}
 	
 	public function test_custom_bits_can_be_coerced_to_snowflakes_or_sonyflakes(): void