+++ CHANGELOG.md
@@ -5,6 +5,20 @@ format. This project adheres to [Semantic Versioning](https://semver.org/spec/v2
 
 ## [Unreleased]
 
+## [0.4.1] - 2024-03-25
+
+## [0.4.0] - 2024-03-12
+
+## [0.3.0] - 2024-02-15
+
+### Added
+
+-   Added a `snowflake_id()` helper method to avoid `Snowflake::make()->id()` calls everywhere
+
+### Changed
+
+-   Improved exception message when trying to set epoch to a future date
+
 ## [0.2.0] - 2023-11-13
 
 ## [0.1.1] - 2023-07-09
@@ -30,7 +44,13 @@ format. This project adheres to [Semantic Versioning](https://semver.org/spec/v2
 -   `Fixed` for any bug fixes. 
 -   `Security` in case of vulnerabilities.
 
-[Unreleased]: https://github.com/glhd/bits/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/glhd/bits/compare/0.4.1...HEAD
+
+[0.4.1]: https://github.com/glhd/bits/compare/0.4.0...0.4.1
+
+[0.4.0]: https://github.com/glhd/bits/compare/0.3.0...0.4.0
+
+[0.3.0]: https://github.com/glhd/bits/compare/0.2.0...0.3.0
 
 [0.2.0]: https://github.com/glhd/bits/compare/0.1.1...0.2.0